perf(contracts): isolate Drive schemas from web sessions
